feat: apply i18n with next-intl
