+++ app/[locale]/(blog)/blog/[slug]/page.tsx
