+++ app/[locale]/(blog)/blog/page.tsx
