+++ app/[locale]/(project)/project/[name]/page.tsx
