+++ i18n/request.ts
@@ -1,0 +1,16 @@
+import {getRequestConfig} from 'next-intl/server';
+import {hasLocale} from 'next-intl';
+import {routing} from './routing';
+
+export default getRequestConfig(async ({requestLocale}) => {
+	// Typically corresponds to the `[locale]` segment
+	const requested = await requestLocale;
+	const locale = hasLocale(routing.locales, requested)
+		? requested
+		: routing.defaultLocale;
+
+	return {
+		locale,
+		messages: (await import(`./${locale}.json`)).default
+	};
+});
