+++ package.json
@@ -23,6 +23,7 @@
 		"clsx": "^2.1.1",
 		"mdast-util-mdxjs-esm": "^2.0.1",
 		"next": "15.2.4",
+		"next-intl": "^4.1.0",
 		"next-mdx-remote": "^5.0.0",
 		"react": "^19.0.0",
 		"react-dom": "^19.0.0",
